pyproject.toml
[project]
name = "rpa-odoo-mesh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "selenium>=4.32.0",
    "tqdm>=4.67.1",
]
